Updated the AI code! Currently using the mobility heuristics as the evaluation function, but still no console output functionality.

+++ src/main/MoveAndValue.java
@@ -3,11 +3,11 @@ package main;
 import com.eudycontreras.othello.capsules.AgentMove;
 
 public class MoveAndValue {
-    public int value;
+    public double value;
     public AgentMove move;
 
 
-    public MoveAndValue(int value, AgentMove move){
+    public MoveAndValue(double value, AgentMove move){
         this.move = move;
         this.value = value;
     }
